Move inspector into own file
